+++ CHANGES.md
@@ -4,6 +4,7 @@
 ## Version 1.2
 
 - new input and output formats *pubtator* and *pubtator_fbk*
+- new normalization method *mask*
 - new parameters `byte-offsets-in` and `byte-offsets-out` for interpreting and producing BioC documents with byte-based offsets
 - fix a security hole involving eval()
 - allow specifying a pretrained sentence splitter